chore: moving NewTxCmd to controller/cli.go
