+++ modules/apps/27-interchain-accounts/controller/client/cli/cli.go
@@ -1,6 +1,7 @@
 package cli
 
 import (
+	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/spf13/cobra"
 )
 
@@ -19,4 +20,22 @@ func GetQueryCmd() *cobra.Command {
 	)
 
 	return queryCmd
+}
+
+// NewTxCmd creates and returns the tx command
+func NewTxCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:                        "controller",
+		Short:                      "ica controller transactions subcommands",
+		DisableFlagParsing:         true,
+		SuggestionsMinimumDistance: 2,
+		RunE:                       client.ValidateCmd,
+	}
+
+	cmd.AddCommand(
+		newRegisterAccountCmd(),
+		newSubmitTxCmd(),
+	)
+
+	return cmd
 }
