+++ modules/apps/27-interchain-accounts/controller/client/cli/tx.go
@@ -23,24 +23,6 @@ const (
 	flagPacketTimeoutTimestamp = "packet-timeout-timestamp"
 )
 
-// NewTxCmd creates and returns the tx command
-func NewTxCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:                        "controller",
-		Short:                      "ica controller transactions subcommands",
-		DisableFlagParsing:         true,
-		SuggestionsMinimumDistance: 2,
-		RunE:                       client.ValidateCmd,
-	}
-
-	cmd.AddCommand(
-		newRegisterAccountCmd(),
-		newSubmitTxCmd(),
-	)
-
-	return cmd
-}
-
 func newRegisterAccountCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "register [connection-id]",
